+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- include USWDS 3.0 for new apps
+
 ## [0.5.3] - 2022-06-06
 
 - check that server started properly before running anything in `bin/with-server`